+++ bindings/nostr-ffi/src/event/kind.rs
@@ -40,14 +40,9 @@ impl Kind {
         Self { inner: e.into() }
     }
 
-    /// Get kind as 16-bit unsigned number
+    /// Get as 16-bit unsigned integer
     pub fn as_u16(&self) -> u16 {
         self.inner.as_u16()
-    }
-
-    /// Get kind as 64-bit unsigned number
-    pub fn as_u64(&self) -> u64 {
-        self.inner.as_u64()
     }
 
     pub fn as_enum(&self) -> KindEnum {